requirements.txt
# Backend & API
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Bot
python-telegram-bot==20.7

# Auth & Security
pyjwt[crypto]==2.8.0
cryptography==42.0.2

# Redis
redis==5.0.1

# Monitoring & Logging
sentry-sdk[fastapi,redis,sqlalchemy]==1.39.2
prometheus-client==0.19.0
structlog==24.1.0

# Utils
pydantic-settings==2.1.0
jinja2==3.1.3
httpx~=0.25.2
